All dishes functions ready

+++ Backend/src/components/dish/services/DishService.js
@@ -2,7 +2,8 @@ const DISH = require('../models/Dish')
 
 const GetOne = async (id) =>
 {
-    let dish = await DISH.findOne({ id: id })
+    let dish = await DISH.findOne({ _id: id })
+
     return dish != null ? { status: 201, data: dish } : { status: 404 }
 }
 
@@ -19,11 +20,13 @@ const Create = async (dish) =>
 
 const Update = async (dish) =>
 {
-    await DISH.findByIdAndUpdate(dish)
+    if (await DISH.findOne({ _id: dish._id }) == null) throw new Error()
+    await DISH.findByIdAndUpdate({ _id: dish._id }, dish)
 }
 
 const Delete = async (id) =>
 {
+    if (await DISH.findOne({ _id: id }) == null) throw new Error()
     await DISH.findByIdAndDelete(id)
 }
